split testfiles
